chore: move backwards compatibility components to their own directory and add readme explaining purpose of custom router context and hoc. use absolute imports
