+++ apps/src/code-studio/legacyDashboardRoutingCompatibility/WithRouterProps.tsx
