+++ apps/src/code-studio/legacyDashboardRoutingCompatibility/index.ts
@@ -1,0 +1,2 @@
+export * from './RouterContext';
+export * from './WithRouterProps';
