+++ apps/src/code-studio/pd/workshop_dashboard/components/workshop_table.jsx
@@ -8,10 +8,10 @@ import {Button} from 'react-bootstrap'; // eslint-disable-line no-restricted-imp
 import * as Table from 'reactabular-table';
 import * as sort from 'sortabular';
 
+import {RouterContext} from '@cdo/apps/code-studio/legacyDashboardRoutingCompatibility';
 import wrappedSortable from '@cdo/apps/templates/tables/wrapped_sortable';
 import color from '@cdo/apps/util/color';
 
-import {RouterContext} from '../../RouterContext';
 import {workshopShape} from '../types.js';
 import {shouldShowSurveyResults} from '../workshop_summary_utils';
