+++ apps/src/code-studio/pd/workshop_dashboard/reports/workshop_summary_report.jsx
@@ -6,8 +6,9 @@ import React from 'react';
 import {Checkbox, Button} from 'react-bootstrap'; // eslint-disable-line no-restricted-imports
 import {connect} from 'react-redux';
 
+import {RouterContext} from '@cdo/apps/code-studio/legacyDashboardRoutingCompatibility';
+
 import Spinner from '../../../../sharedComponents/Spinner';
-import {RouterContext} from '../../RouterContext';
 import {PermissionPropType, WorkshopAdmin} from '../permission';
 
 import {QUERY_BY_VALUES, COURSE_VALUES} from './report_constants';
